Refactored for multiple inheritance.
This is a big change, breaking for a few things. Particularly:
 * atom no longer has get-value but has get-atom-value
 * there is no longer a separate boolean with get-value; instead, atom can be
   used as an alias for boolean, but that will require calling
   get-boolean-value
 * list no longer inherits from tuple
 * string no longer inherits from list
 * there is no longer a sparate char namespace (moved into int)

In summary, multiple object inheritance is now handled in a functional
programmatic manner: multiple calls to (extend ...) with different prototcols
(but the same JInterface object). This results in much cleaner code (less
duplication, more FP) and in amore Clojure-idiomatic manner.